try to make the batch loaders stuff comprehensible
by splitting out inner classes and adding javadoc
